WIP on incorporating the new CLI properly
