+++ composeApp/src/commonMain/kotlin/de/medizininformatikinitiative/medgraph/ui/Main.kt
@@ -8,13 +8,9 @@ import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
 
-fun main(args: Array<String>) {
+fun main() {
     LogManager.getLogger("Application").log(Level.INFO, "Application started.")
-
-    val cmdLineExecutor = CommandLineExecutor();
-    val exitStatus = cmdLineExecutor.evaluateAndExecuteCommandLineArguments(args)
-    if (exitStatus.isPresent) System.exit(exitStatus.asInt)
-    else UI.startUi(!hasFunctioningDatabaseConnection())
+    UI.startUi(!hasFunctioningDatabaseConnection())
 }
 
 private fun hasFunctioningDatabaseConnection() =
