+++ composeApp/src/desktopMain/kotlin/de/medizininformatikinitiative/medgraph/ui/desktop/Main.kt
@@ -1,9 +1,14 @@
 package de.medizininformatikinitiative.medgraph.ui.desktop
 
+import de.medizininformatikinitiative.medgraph.commandline.CommandLineExecutor
 import de.medizininformatikinitiative.medgraph.common.logging.LogManager
 import de.medizininformatikinitiative.medgraph.ui.desktop.logging.Log4J2LoggerFactory
 
-fun main() {
+fun main(args: Array<String>) {
     LogManager.setLogFactory(Log4J2LoggerFactory())
-    de.medizininformatikinitiative.medgraph.ui.main()
+
+    val cmdLineExecutor = CommandLineExecutor();
+    val exitStatus = cmdLineExecutor.evaluateAndExecuteCommandLineArguments(args)
+    if (exitStatus.isPresent) System.exit(exitStatus.asInt)
+    else de.medizininformatikinitiative.medgraph.ui.main()
 }
